backend/requirements.txt
# SentinelVision backend
fastapi>=0.109.0,<1.0
uvicorn[standard]>=0.27.0,<1.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ML
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.36.0
Pillow>=10.0.0
accelerate>=0.25.0
